Re-structured for final submission
Worked with @JamesShan67 to implement loading bars and the playlist view page. Also cleaned up repo for final submission
