+++ client/package.json
@@ -12,11 +12,16 @@
     "@types/react": "^18.0.38",
     "@types/react-dom": "^18.0.11",
     "framer-motion": "^10.12.4",
+    "loaders.css": "^0.1.2",
     "mdb-react-ui-kit": "^6.0.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-loader-spinner": "^5.3.4",
+    "react-loaders": "^3.0.1",
     "react-router-dom": "^6.10.0",
     "react-scripts": "5.0.1",
+    "react-spinner-animated": "^3.1.1",
+    "react-spinners": "^0.13.8",
     "typescript": "^4.9.5",
     "web-vitals": "^2.1.4"
   },
