+++ client/src/pages/HomePage.tsx
@@ -1,4 +1,3 @@
-
 import React, {useState, useEffect} from 'react';
 import httpClient from "../httpClient";
 import { User } from '../types'
@@ -68,10 +67,8 @@ useEffect(() => {
         </h1>
 
         <p className='px-3' style={{color: 'hsl(218, 81%, 85%)'}}>
-          Lorem ipsum dolor sit amet consectetur adipisicing elit.
-          Eveniet, itaque accusantium odio, soluta, corrupti aliquam
-          quibusdam tempora at cupiditate quis eum maiores libero
-          veritatis? Dicta facilis sint aliquid ipsum atque?
+        Create your own unique playlist and discover new music with our web application. \n
+        Login with Spotify, answer a few questions about your musical preferences, and enjoy a custom playlist tailored just for you.
         </p>
 
       </MDBCol>
