+++ documentation/prototype_apis/dalle.py
@@ -2,14 +2,17 @@ import openai
 import os
 import requests
 import base64
+from dotenv import load_dotenv
+
+load_dotenv()
 
 # Set up the OpenAI API client
-openai.api_key = ""
+openai.api_key = os.getenv("gpt_key")
 
-songstyle = input("Enter the your song style")
+
 
 # Define the input prompt for DALL-E
-userprompt = "Generate a playlist cover with songs in the style of " + songstyle
+userprompt = "Rustic old garage with yellowish stains with an old car parked in front of it with eggs" 
 
 # Define the DALL-E API endpoint
 endpoint = "https://api.openai.com/v1/images/generations"
@@ -19,7 +22,7 @@ data = {
     "model": "image-alpha-001",
     "prompt": userprompt,
     "num_images": 1,
-    "size": "1024x1024",
+    "size": "512x512",
     "response_format": "url"
 }
 
@@ -35,5 +38,8 @@ image_data = requests.get(image_url).content
 # Convert the image data to a base64-encoded string
 image_base64 = base64.b64encode(image_data).decode("utf-8")
 
-# Print the base64-encoded image data
-print(image_base64)
+# # Print the base64-encoded image data
+#image_resp = openai.Image.create(prompt="Generate an artistic album cover with songs like Midnight City By M83, Solitude is Bliss By Tame Impala, and Inro by the xx and without text", n=1, size="512x512")
+
+models = openai.Model.list()
+print(image_url)
