+++ flask-server/.cache
@@ -1,0 +1,1 @@
+{"access_token": "BQAa2LjoMIGDwPxIqtctiTTAYoijU7tHLUoDFPjdqtOZe6ywQC2Yf6BT_whqa8q0z2bKv5RZYvzhXuBs0grPxOhLM0qenIIARyS9zJGensE7oAPlgqPG_8GOhDhkQuyfbeGk_38nw2RQeZVnbmpgqXqExlv3LftrjWfRoAQT0MRu6zlF2dd92SWVzoxSidaDiVc2HxLHSoYG1BkwBiiASdZ5BD-r2OxbEPToZkA-euo", "token_type": "Bearer", "expires_in": 3600, "scope": "playlist-modify-private ugc-image-upload user-library-modify user-library-read", "expires_at": 1683061574, "refresh_token": "AQCP9IHs0KP4Bi1PCJllZYdS3rhvx9dOv3PcEjdYpnKrkHVRoxSvwBkMvBedsjP5rDWMWdwFNiHwL5HtXXarxwiaGuEbijv9eMqXRhrGBspm0AavjtjsFnhVatEHk8ITAzY"}
